+++ src/digitalIdentity-frontend/src/app/app.module.ts
@@ -3,27 +3,27 @@ import { BrowserModule } from '@angular/platform-browser';
 
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
-import { LoginPageComponent } from './login-page/login-page.component';
-import { CreateDIPageComponent } from './createDI-page/createDI-page.component';
-import { HomeComponent } from './home/home.component';
-import { DIOverviewComponent } from './DI-Overview/DI-Overview.component';
-import { SettingsComponent } from './settings/settings.component';
-import { InformationPopUpComponent } from './information-pop-up/information-pop-up.component';
-import { EditWindowPopUpComponent } from './edit-window-pop-up/edit-window-pop-up.component';
-import { NavigationBarComponent } from './navigation-bar/navigation-bar.component';
-import { SchemaOverviewComponent } from './schema-overview/schema-overview.component';
+import { LoginPageComponent } from './components/login-page/login-page.component';
+import { CreateDIPageComponent } from './components/createDI-page/createDI-page.component';
+import { HomeComponent } from './components/home/home.component';
+import { DIOverviewComponent } from './components/DI-Overview/DI-Overview.component';
+import { SettingsComponent } from './components/settings/settings.component';
+import { InformationPopUpComponent } from './components/information-pop-up/information-pop-up.component';
+import { EditWindowPopUpComponent } from './components/edit-window-pop-up/edit-window-pop-up.component';
+import { NavigationBarComponent } from './components/navigation-bar/navigation-bar.component';
+import { SchemaOverviewComponent } from './components/schema-overview/schema-overview.component';
 // import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
 
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 // imports the MatModule: a module which loads contains all necessary @angular/material/ imports
-import { MaterialModule } from './material/material.module';
+import { MaterialModule } from './components/material/material.module';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
-import { CreateSchemaPageComponent } from './create-schema-page/create-schema-page.component';
-import { ShowSchemaPopUpComponent } from './show-schema-pop-up/show-schema-pop-up.component';
-import { ErrorPageComponent } from './error-pages/error-page/error-page.component';
-import { MenuItemComponent } from './menu-item/menu-item.component';
+import { CreateSchemaPageComponent } from './components/create-schema-page/create-schema-page.component';
+import { ShowSchemaPopUpComponent } from './components/show-schema-pop-up/show-schema-pop-up.component';
+import { ErrorPageComponent } from './components/error-page/error-page.component';
+import { MenuItemComponent } from './components/menu-item/menu-item.component';
 
 @NgModule({
   declarations: [
