+++ src/digitalIdentity-frontend/src/app/components/create-schema-page/create-schema-page.component.spec.ts
