+++ src/digitalIdentity-frontend/src/app/components/error-page/error-page.component.css
