+++ src/digitalIdentity-frontend/src/app/components/information-pop-up/information-pop-up.component.spec.ts
