+++ src/digitalIdentity-frontend/src/app/components/menu-item/menu-item.component.ts
