+++ src/digitalIdentity-frontend/src/app/components/navigation-bar/navigation-bar.component.css
