+++ src/digitalIdentity-frontend/src/app/components/navigation-bar/navigation-bar.component.spec.ts
