+++ src/digitalIdentity-frontend/src/app/components/schema-overview/schema-overview.component.ts
