+++ src/digitalIdentity-frontend/src/app/components/settings/settings.component.spec.ts
