+++ src/digitalIdentity-frontend/src/app/components/show-schema-pop-up/show-schema-pop-up.component.spec.ts
