+++ src/digitalIdentity-frontend/src/app/services/backend-http-service/backend-http-service.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { BackendHttpService} from './backend-http-service.service';
+
+describe('BackendHttpServiceService', () => {
+  let service: BackendHttpService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(BackendHttpService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
