+++ src/digitalIdentity-frontend/src/app/services/backend-http-service/backend-http-service.service.ts
@@ -1,0 +1,61 @@
+import { HttpClient, HttpErrorResponse, HttpHeaders, HttpParams, HttpResponse } from '@angular/common/http';
+import { Injectable, isDevMode } from '@angular/core';
+import { catchError, firstValueFrom, Observable, of, timeout } from 'rxjs';
+import { environment } from 'src/environments/environment';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class BackendHttpService {
+  headers = new HttpHeaders().append(
+    'Content-Type',
+    'application/json'
+  );
+  constructor(private http: HttpClient) { }
+
+  async postRequest(processName: string,path:string,data:any, params:HttpParams) : Promise<any> {
+    let body = JSON.stringify(data)
+    let promise = await new Promise<HttpResponse<any>>((resolve,reject) => (<Observable<HttpResponse<any>>>this.http
+    .post<any>(environment.serverURL + '/auth/register', body, {
+      headers: this.headers,
+      observe: 'response',
+      params: params,
+    })
+    .pipe(
+      timeout(2000),
+       // timeout after 2 seconds
+      catchError(e => {
+        return of(e);
+      }
+      )))
+    .subscribe({
+      next: (response) => {
+        if (response.ok && isDevMode()) {
+          console.log(processName + ' successful! Server response:');
+          console.log(response);
+        } else if (isDevMode()) {
+          console.log(processName + ' not successful! Got Error message:');
+          console.log(response);
+        }
+        resolve(response);
+      },
+      error: (error) => {
+        if (isDevMode()) {
+          if (error.name == "TimeoutError") {
+            console.log(processName + " timed out")
+          }
+          else {
+            console.log(processName + ' not successful!: Error: ');
+            console.log(error);
+          }
+        }
+        reject(error);
+      },
+    }));
+    return promise
+  }
+
+  getRequest(processName: string) {
+
+  }
+}
